Agresta:  # commit versione develer chiusa per impianti pilota preliminari    # implementazione sorgente Multimediale in Diffusione sonora Classica    # implementazione sorgente Multimediale in Diffusione sonora multicanale    # implementazione centrale 4 zone    # implementazione sonde non controllate    # implementazione sonde esterne    # implemntazione sonde fan-coil  # sistemazioni varie per stampe testi, regolazione sveglia, gestione chi 22

+++ sveglia.h
@@ -186,7 +186,7 @@ private:
    unsigned int contaBuzzer;
    QDateTime* oraSveglia;
    contdiff* difson;
-   uchar	volSveglia[AMPLI_NUM];
+   int	volSveglia[AMPLI_NUM];
    bool	gesFrameAbil,svegliaAbil,onceToGest;
    QTimer   *minuTimer,*aumVolTimer;
    char*	frame;
